Fix reviewer feedback: retry config, mutation keys, security hardening, JSDoc, RSC tests
Address 7 items from PR #6 code review:
- Fix retry config to use actual API error codes (PERMISSION_DENIED instead of FORBIDDEN)
- Add mutation key factory (tripKeys.create/update/cancel)
- Narrow images.remotePatterns from wildcard to localhost
- Replace hardcoded tripful.com with NEXT_PUBLIC_SITE_URL env var
- Remove error.message exposure in global-error.tsx
- Add per-property JSDoc to shared/types/trip.ts
- Add RSC page-level tests for dashboard and trip detail prefetch pipeline
