+++ apps/web/.env.local.example
@@ -31,6 +31,14 @@ NEXT_PUBLIC_API_URL=http://localhost:8000/api
 API_URL=http://localhost:8000/api
 
 # ===========================================
+# Site URL Configuration
+# ===========================================
+
+# Public site URL used for SEO (robots.txt, sitemap.xml)
+# In production, set to your actual domain (e.g., https://tripful.com)
+NEXT_PUBLIC_SITE_URL=http://localhost:3000
+
+# ===========================================
 # Optional: Additional Frontend Configuration
 # ===========================================
