+++ apps/web/src/app/(app)/trips/[id]/page.test.tsx
@@ -1,985 +1,126 @@
-import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
-import {
-  render,
-  screen,
-  waitFor,
-  act,
-  fireEvent,
-} from "@testing-library/react";
-import userEvent from "@testing-library/user-event";
-import { Suspense } from "react";
-import { TripDetailContent } from "./trip-detail-content";
-import type { TripDetail } from "@/hooks/use-trips";
-import type { User } from "@tripful/shared";
+import { describe, it, expect, vi, beforeEach } from "vitest";
+import { render } from "@testing-library/react";
 
-// Mock next/dynamic
-vi.mock("next/dynamic", () => ({
-  __esModule: true,
-  default: (importFn: () => Promise<any>) => {
-    const React = require("react");
-    const Lazy = React.lazy(importFn);
-    return function DynamicComponent(props: any) {
-      return React.createElement(
-        React.Suspense,
-        { fallback: null },
-        React.createElement(Lazy, props),
-      );
-    };
-  },
+// Mock server API
+const mockServerApiRequest = vi.fn();
+vi.mock("@/lib/server-api", () => ({
+  serverApiRequest: (...args: unknown[]) => mockServerApiRequest(...args),
 }));
 
-// Mock next/image
-vi.mock("next/image", () => ({
-  default: ({
-    src,
-    alt,
-    fill,
-    priority,
-    unoptimized,
-    sizes,
-    ...props
-  }: Record<string, unknown>) => (
-    // eslint-disable-next-line @next/next/no-img-element, jsx-a11y/alt-text
-    <img src={src as string} alt={alt as string} {...props} />
-  ),
-}));
-
-// Mock next/link
-vi.mock("next/link", () => ({
-  default: ({
-    href,
-    children,
-    ...props
-  }: {
-    href: string;
-    children: React.ReactNode;
-    [key: string]: unknown;
-  }) => (
-    <a href={href} {...props}>
-      {children}
-    </a>
-  ),
-}));
-
-// Mock next/navigation
-const mockPush = vi.fn();
-vi.mock("next/navigation", () => ({
-  useRouter: () => ({
-    push: mockPush,
+// Mock query client
+const mockSetQueryData = vi.fn();
+vi.mock("@/lib/get-query-client", () => ({
+  getQueryClient: () => ({
+    setQueryData: mockSetQueryData,
   }),
 }));
 
-// Mock useAuth hook
-const mockUser: User = {
-  id: "user-123",
-  phoneNumber: "+14155551234",
-  displayName: "John Doe",
-  profilePhotoUrl: "https://example.com/john.jpg",
-  timezone: "America/New_York",
-  createdAt: new Date("2026-01-01T00:00:00Z"),
-  updatedAt: new Date("2026-01-01T00:00:00Z"),
-};
+// Mock TanStack Query — preserve real exports (queryOptions used by trip-queries.ts)
+vi.mock("@tanstack/react-query", async (importOriginal) => {
+  const actual = await importOriginal<typeof import("@tanstack/react-query")>();
+  return {
+    ...actual,
+    dehydrate: vi.fn(() => ({ queries: [] })),
+    HydrationBoundary: ({ children }: { children: React.ReactNode }) => (
+      <div data-testid="hydration-boundary">{children}</div>
+    ),
+  };
+});
 
-const mockUseAuth = vi.fn();
-vi.mock("@/app/providers/auth-provider", () => ({
-  useAuth: () => mockUseAuth(),
-}));
-
-// Mock useTripDetail hook
-const mockUseTripDetail = vi.fn();
-vi.mock("@/hooks/use-trips", () => ({
-  useTripDetail: (tripId: string) => mockUseTripDetail(tripId),
-}));
-
-// Mock EditTripDialog component
-vi.mock("@/components/trip/edit-trip-dialog", () => ({
-  EditTripDialog: ({
-    open,
-    onOpenChange,
-    trip,
-    onSuccess,
-  }: {
-    open: boolean;
-    onOpenChange: (open: boolean) => void;
-    trip: TripDetail;
-    onSuccess?: () => void;
-  }) => (
-    <div
-      data-testid="edit-trip-dialog"
-      data-open={open}
-      data-onsuccess={onSuccess ? "true" : "false"}
-    >
-      <button onClick={() => onOpenChange(false)}>Close Dialog</button>
-      <span>Editing: {trip.name}</span>
-      {onSuccess && (
-        <button
-          data-testid="trigger-success"
-          onClick={() => {
-            onSuccess();
-            onOpenChange(false);
-          }}
-        >
-          Trigger Success
-        </button>
-      )}
-    </div>
+// Mock TripDetailContent
+vi.mock("./trip-detail-content", () => ({
+  TripDetailContent: ({ tripId }: { tripId: string }) => (
+    <div data-testid="trip-detail-content" data-trip-id={tripId} />
   ),
 }));
 
-describe("TripDetailContent", () => {
-  const mockTripDetail: TripDetail = {
+// Import AFTER mocks
+import TripDetailPage, { generateMetadata } from "./page";
+
+const mockTripResponse = {
+  success: true as const,
+  trip: {
     id: "trip-123",
-    name: "Bachelor Party in Miami",
-    destination: "Miami Beach, FL",
-    startDate: "2026-06-01",
-    endDate: "2026-06-05",
-    preferredTimezone: "America/New_York",
-    description: "Epic bachelor party weekend with the crew!",
-    coverImageUrl: "https://example.com/cover.jpg",
-    createdBy: "user-123",
+    name: "Beach Trip",
+    destination: "Hawaii",
+    startDate: null,
+    endDate: null,
+    preferredTimezone: "Pacific/Honolulu",
+    description: null,
+    coverImageUrl: null,
+    createdBy: "user-1",
     allowMembersToAddEvents: true,
     cancelled: false,
-    createdAt: new Date("2026-02-06T12:00:00Z"),
-    updatedAt: new Date("2026-02-06T12:00:00Z"),
-    organizers: [
-      {
-        id: "user-123",
-        displayName: "John Doe",
-        phoneNumber: "+14155551234",
-        profilePhotoUrl: "https://example.com/john.jpg",
-        timezone: "America/New_York",
-      },
-      {
-        id: "user-456",
-        displayName: "Jane Smith",
-        phoneNumber: "+14155555678",
-        profilePhotoUrl: null,
-        timezone: "America/Los_Angeles",
-      },
-    ],
-    memberCount: 8,
-  };
+    createdAt: new Date(),
+    updatedAt: new Date(),
+    organizers: [],
+    memberCount: 1,
+  },
+};
 
+describe("TripDetailPage (RSC)", () => {
   beforeEach(() => {
     vi.clearAllMocks();
-    mockUseAuth.mockReturnValue({ user: mockUser });
   });
 
-  afterEach(() => {
-    vi.useRealTimers();
+  it("awaits params and prefetches trip detail", async () => {
+    mockServerApiRequest.mockResolvedValue(mockTripResponse);
+
+    const result = await TripDetailPage({
+      params: Promise.resolve({ id: "trip-123" }),
+    });
+    render(result as React.ReactElement);
+
+    expect(mockServerApiRequest).toHaveBeenCalledWith("/trips/trip-123");
+    expect(mockSetQueryData).toHaveBeenCalledWith(
+      ["trips", "trip-123"],
+      mockTripResponse.trip,
+    );
   });
 
-  describe("rendering tests", () => {
-    it("renders loading state initially", () => {
-      mockUseTripDetail.mockReturnValue({
-        data: undefined,
-        isPending: true,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
+  it("renders HydrationBoundary with TripDetailContent even when prefetch fails", async () => {
+    mockServerApiRequest.mockRejectedValue(new Error("Server error"));
 
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      // Check for skeleton loading state
-      const skeletonElements = document.querySelectorAll(".animate-pulse");
-      expect(skeletonElements.length).toBeGreaterThan(0);
+    const result = await TripDetailPage({
+      params: Promise.resolve({ id: "trip-456" }),
     });
+    const { getByTestId } = render(result as React.ReactElement);
 
-    it("renders trip details when data loaded", async () => {
-      mockUseTripDetail.mockReturnValue({
-        data: mockTripDetail,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("Bachelor Party in Miami")).toBeDefined();
-      });
-
-      expect(screen.getByText("Miami Beach, FL")).toBeDefined();
-      expect(screen.getByText("Jun 1 - 5, 2026")).toBeDefined();
-      expect(screen.getByText("8 members")).toBeDefined();
-      expect(screen.getByText("0 events")).toBeDefined();
-    });
-
-    it("shows cover image when available", async () => {
-      mockUseTripDetail.mockReturnValue({
-        data: mockTripDetail,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        const coverImage = screen.getByAltText("Bachelor Party in Miami");
-        expect(coverImage).toBeDefined();
-        expect(coverImage.getAttribute("src")).toBe(
-          "https://example.com/cover.jpg",
-        );
-      });
-    });
-
-    it("shows placeholder when no cover image", async () => {
-      const tripWithoutCover = { ...mockTripDetail, coverImageUrl: null };
-      mockUseTripDetail.mockReturnValue({
-        data: tripWithoutCover,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("Bachelor Party in Miami")).toBeDefined();
-      });
-
-      // Check for gradient placeholder
-      const placeholder = document.querySelector(
-        ".bg-gradient-to-br.from-slate-100.to-blue-100",
-      );
-      expect(placeholder).toBeDefined();
-    });
-
-    it("displays trip name, destination, dates", async () => {
-      mockUseTripDetail.mockReturnValue({
-        data: mockTripDetail,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("Bachelor Party in Miami")).toBeDefined();
-        expect(screen.getByText("Miami Beach, FL")).toBeDefined();
-        expect(screen.getByText("Jun 1 - 5, 2026")).toBeDefined();
-      });
-    });
-
-    it("shows organizer information", async () => {
-      mockUseTripDetail.mockReturnValue({
-        data: mockTripDetail,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("Organizers")).toBeDefined();
-        expect(screen.getByText("John Doe, Jane Smith")).toBeDefined();
-      });
-
-      // Check for organizer avatars
-      const johnAvatar = screen.getByAltText("John Doe");
-      expect(johnAvatar).toBeDefined();
-      expect(johnAvatar.getAttribute("src")).toBe(
-        "https://example.com/john.jpg",
-      );
-
-      // Jane Smith has no profile photo, should show initials
-      expect(screen.getByText("JS")).toBeDefined();
-    });
-
-    it("displays member count and event count", async () => {
-      mockUseTripDetail.mockReturnValue({
-        data: mockTripDetail,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("8 members")).toBeDefined();
-        expect(screen.getByText("0 events")).toBeDefined();
-      });
-    });
-
-    it("shows RSVP badge", async () => {
-      mockUseTripDetail.mockReturnValue({
-        data: mockTripDetail,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("Going")).toBeDefined();
-      });
-    });
-
-    it("shows description when available", async () => {
-      mockUseTripDetail.mockReturnValue({
-        data: mockTripDetail,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("About this trip")).toBeDefined();
-        expect(
-          screen.getByText("Epic bachelor party weekend with the crew!"),
-        ).toBeDefined();
-      });
-    });
-
-    it("hides description section when description is null", async () => {
-      const tripWithoutDescription = { ...mockTripDetail, description: null };
-      mockUseTripDetail.mockReturnValue({
-        data: tripWithoutDescription,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("Bachelor Party in Miami")).toBeDefined();
-      });
-
-      expect(screen.queryByText("About this trip")).toBeNull();
-    });
+    expect(getByTestId("hydration-boundary")).toBeDefined();
+    expect(getByTestId("trip-detail-content")).toBeDefined();
+    expect(
+      getByTestId("trip-detail-content").getAttribute("data-trip-id"),
+    ).toBe("trip-456");
+    expect(mockSetQueryData).not.toHaveBeenCalled();
   });
 
-  describe("authorization tests", () => {
-    it("shows edit button when user is creator", async () => {
-      mockUseAuth.mockReturnValue({ user: mockUser });
-      mockUseTripDetail.mockReturnValue({
-        data: mockTripDetail,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
+  it("does not throw when prefetch fails", async () => {
+    mockServerApiRequest.mockRejectedValue(new Error("Network error"));
 
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("Edit trip")).toBeDefined();
-      });
-    });
-
-    it("shows edit button when user is co-organizer", async () => {
-      const coOrganizerUser = { ...mockUser, id: "user-456" };
-      mockUseAuth.mockReturnValue({ user: coOrganizerUser });
-      mockUseTripDetail.mockReturnValue({
-        data: mockTripDetail,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("Edit trip")).toBeDefined();
-      });
-    });
-
-    it("hides edit button when user is not organizer", async () => {
-      const regularUser = { ...mockUser, id: "user-789" };
-      mockUseAuth.mockReturnValue({ user: regularUser });
-      mockUseTripDetail.mockReturnValue({
-        data: mockTripDetail,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("Bachelor Party in Miami")).toBeDefined();
-      });
-
-      expect(screen.queryByText("Edit trip")).toBeNull();
-    });
-
-    it("shows organizer badge for organizers only", async () => {
-      mockUseAuth.mockReturnValue({ user: mockUser });
-      mockUseTripDetail.mockReturnValue({
-        data: mockTripDetail,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("Organizing")).toBeDefined();
-      });
-    });
-
-    it("hides organizer badge for non-organizers", async () => {
-      const regularUser = { ...mockUser, id: "user-789" };
-      mockUseAuth.mockReturnValue({ user: regularUser });
-      mockUseTripDetail.mockReturnValue({
-        data: mockTripDetail,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("Bachelor Party in Miami")).toBeDefined();
-      });
-
-      expect(screen.queryByText("Organizing")).toBeNull();
-    });
+    await expect(
+      TripDetailPage({ params: Promise.resolve({ id: "trip-789" }) }),
+    ).resolves.not.toThrow();
   });
 
-  describe("error handling tests", () => {
-    it("shows error state on 404", async () => {
-      mockUseTripDetail.mockReturnValue({
-        data: undefined,
-        isPending: false,
-        isError: true,
-        error: new Error("NOT_FOUND"),
-        refetch: vi.fn(),
-      });
+  it("passes tripId from params to TripDetailContent", async () => {
+    mockServerApiRequest.mockResolvedValue(mockTripResponse);
 
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("Trip not found")).toBeDefined();
-      });
-
-      expect(
-        screen.getByText(
-          "This trip doesn't exist or you don't have access to it.",
-        ),
-      ).toBeDefined();
+    const result = await TripDetailPage({
+      params: Promise.resolve({ id: "trip-abc" }),
     });
+    const { getByTestId } = render(result as React.ReactElement);
 
-    it("shows error state on network failure", async () => {
-      mockUseTripDetail.mockReturnValue({
-        data: undefined,
-        isPending: false,
-        isError: true,
-        error: new Error("Network error"),
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("Trip not found")).toBeDefined();
-      });
-    });
-
-    it("return to dashboard link has correct href", async () => {
-      mockUseTripDetail.mockReturnValue({
-        data: undefined,
-        isPending: false,
-        isError: true,
-        error: new Error("NOT_FOUND"),
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("Trip not found")).toBeDefined();
-      });
-
-      const returnLink = screen.getByText("Return to dashboard");
-      expect(returnLink.closest("a")?.getAttribute("href")).toBe("/dashboard");
-    });
+    expect(
+      getByTestId("trip-detail-content").getAttribute("data-trip-id"),
+    ).toBe("trip-abc");
   });
+});
 
-  describe("interaction tests", () => {
-    it("edit button opens EditTripDialog", async () => {
-      mockUseAuth.mockReturnValue({ user: mockUser });
-      mockUseTripDetail.mockReturnValue({
-        data: mockTripDetail,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      const user = userEvent.setup();
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("Edit trip")).toBeDefined();
-      });
-
-      const editButton = screen.getByText("Edit trip");
-      await user.click(editButton);
-
-      await waitFor(() => {
-        const dialog = screen.getByTestId("edit-trip-dialog");
-        expect(dialog).toBeDefined();
-        expect(dialog.getAttribute("data-open")).toBe("true");
-      });
+describe("generateMetadata", () => {
+  it("returns title with trip ID", async () => {
+    const metadata = await generateMetadata({
+      params: Promise.resolve({ id: "trip-123" }),
     });
-
-    it("can close EditTripDialog", async () => {
-      mockUseAuth.mockReturnValue({ user: mockUser });
-      mockUseTripDetail.mockReturnValue({
-        data: mockTripDetail,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      const user = userEvent.setup();
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("Edit trip")).toBeDefined();
-      });
-
-      // Open dialog
-      const editButton = screen.getByText("Edit trip");
-      await user.click(editButton);
-
-      await waitFor(() => {
-        expect(
-          screen.getByTestId("edit-trip-dialog").getAttribute("data-open"),
-        ).toBe("true");
-      });
-
-      // Close dialog
-      const closeButton = screen.getByText("Close Dialog");
-      await user.click(closeButton);
-
-      await waitFor(() => {
-        expect(
-          screen.getByTestId("edit-trip-dialog").getAttribute("data-open"),
-        ).toBe("false");
-      });
-    });
-  });
-
-  describe("success notification tests", () => {
-    it("shows success banner when trip is updated", async () => {
-      mockUseAuth.mockReturnValue({ user: mockUser });
-      mockUseTripDetail.mockReturnValue({
-        data: mockTripDetail,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("Edit trip")).toBeDefined();
-      });
-
-      // Success banner should not be visible initially
-      expect(screen.queryByText(/trip updated successfully/i)).toBeNull();
-
-      // Trigger onSuccess callback by simulating trip update
-      // The mock EditTripDialog component will have an onSuccess prop
-      const dialog = screen.getByTestId("edit-trip-dialog");
-      const onSuccessProp = dialog.getAttribute("data-onsuccess");
-      expect(onSuccessProp).toBe("true");
-    });
-
-    it("success banner auto-dismisses after timeout", () => {
-      vi.useFakeTimers();
-
-      mockUseAuth.mockReturnValue({ user: mockUser });
-      mockUseTripDetail.mockReturnValue({
-        data: mockTripDetail,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      // Open dialog
-      const editButton = screen.getByText("Edit trip");
-      fireEvent.click(editButton);
-
-      // Trigger success by clicking the success trigger button (added to mock)
-      const successButton = screen.getByTestId("trigger-success");
-      fireEvent.click(successButton);
-
-      // Success banner should appear
-      expect(screen.getByText(/trip updated successfully/i)).toBeDefined();
-
-      // Fast-forward time by 5 seconds
-      act(() => {
-        vi.advanceTimersByTime(5000);
-      });
-
-      // Banner should now be dismissed
-      expect(screen.queryByText(/trip updated successfully/i)).toBeNull();
-
-      vi.useRealTimers();
-    });
-
-    it("success banner has correct styling", async () => {
-      mockUseAuth.mockReturnValue({ user: mockUser });
-      mockUseTripDetail.mockReturnValue({
-        data: mockTripDetail,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      const user = userEvent.setup();
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      // Open dialog
-      const editButton = screen.getByText("Edit trip");
-      await user.click(editButton);
-
-      // Trigger success
-      const successButton = screen.getByTestId("trigger-success");
-      await user.click(successButton);
-
-      // Check banner styling
-      const banner = screen.getByText(
-        /trip updated successfully/i,
-      ).parentElement;
-      expect(banner).toBeDefined();
-      expect(banner?.className).toContain("bg-green-50");
-      expect(banner?.className).toContain("border-green-200");
-    });
-
-    it("does not show success banner on initial page load", () => {
-      mockUseAuth.mockReturnValue({ user: mockUser });
-      mockUseTripDetail.mockReturnValue({
-        data: mockTripDetail,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      expect(screen.queryByText(/trip updated successfully/i)).toBeNull();
-    });
-  });
-
-  describe("edge cases", () => {
-    it("handles missing dates (shows Dates TBD)", async () => {
-      const tripWithoutDates = {
-        ...mockTripDetail,
-        startDate: null,
-        endDate: null,
-      };
-      mockUseTripDetail.mockReturnValue({
-        data: tripWithoutDates,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("Dates TBD")).toBeDefined();
-      });
-    });
-
-    it("handles null description (doesn't render section)", async () => {
-      const tripWithoutDescription = { ...mockTripDetail, description: null };
-      mockUseTripDetail.mockReturnValue({
-        data: tripWithoutDescription,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("Bachelor Party in Miami")).toBeDefined();
-      });
-
-      expect(screen.queryByText("About this trip")).toBeNull();
-    });
-
-    it("handles empty organizers list", async () => {
-      const tripWithoutOrganizers = { ...mockTripDetail, organizers: [] };
-      mockUseTripDetail.mockReturnValue({
-        data: tripWithoutOrganizers,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("Bachelor Party in Miami")).toBeDefined();
-      });
-
-      // Organizers section should not be rendered
-      expect(screen.queryByText("Organizers")).toBeNull();
-    });
-
-    it("handles 0 members", async () => {
-      const tripWithNoMembers = { ...mockTripDetail, memberCount: 0 };
-      mockUseTripDetail.mockReturnValue({
-        data: tripWithNoMembers,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("0 members")).toBeDefined();
-      });
-    });
-
-    it("handles 1 member (singular)", async () => {
-      const tripWithOneMember = { ...mockTripDetail, memberCount: 1 };
-      mockUseTripDetail.mockReturnValue({
-        data: tripWithOneMember,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("1 member")).toBeDefined();
-      });
-    });
-
-    it("handles dates spanning different months", async () => {
-      const tripCrossMonth = {
-        ...mockTripDetail,
-        startDate: "2026-06-28",
-        endDate: "2026-07-05",
-      };
-      mockUseTripDetail.mockReturnValue({
-        data: tripCrossMonth,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("Jun 28 - Jul 5, 2026")).toBeDefined();
-      });
-    });
-
-    it("handles only start date", async () => {
-      const tripOnlyStart = {
-        ...mockTripDetail,
-        startDate: "2026-06-01",
-        endDate: null,
-      };
-      mockUseTripDetail.mockReturnValue({
-        data: tripOnlyStart,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("Starts Jun 1, 2026")).toBeDefined();
-      });
-    });
-
-    it("handles only end date", async () => {
-      const tripOnlyEnd = {
-        ...mockTripDetail,
-        startDate: null,
-        endDate: "2026-06-05",
-      };
-      mockUseTripDetail.mockReturnValue({
-        data: tripOnlyEnd,
-        isPending: false,
-        isError: false,
-        error: null,
-        refetch: vi.fn(),
-      });
-
-      render(
-        <Suspense fallback={null}>
-          <TripDetailContent tripId="trip-123" />
-        </Suspense>,
-      );
-
-      await waitFor(() => {
-        expect(screen.getByText("Ends Jun 5, 2026")).toBeDefined();
-      });
-    });
+    expect(metadata).toEqual({ title: "Trip trip-123" });
   });
 });
