+++ apps/web/src/app/robots.ts
@@ -1,12 +1,13 @@
 import type { MetadataRoute } from "next";
 
 export default function robots(): MetadataRoute.Robots {
+  const siteUrl = process.env.NEXT_PUBLIC_SITE_URL || "https://tripful.com";
   return {
     rules: {
       userAgent: "*",
       allow: "/",
       disallow: ["/dashboard", "/trips"],
     },
-    sitemap: "https://tripful.com/sitemap.xml",
+    sitemap: `${siteUrl}/sitemap.xml`,
   };
 }
